Fix test_llm.py: move cases to tests_with_llm
